[update] dummy用のregistryを追加
